+++ CHANGELOG.md
@@ -6,9 +6,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Default seller search to make filters work as well
 
 ## [3.71.3] - 2020-09-09
-
 ### Fixed
 - Feedback on facet selection from the selected facets